new role to add

+++ Articels/Articels/Controllers/Api/DeleteController.cs
@@ -2,6 +2,7 @@
 using Articels.Models;
 using Articels.Models.Repository;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Query.Internal;
@@ -14,19 +15,27 @@ namespace Articels.Controllers.Api
     {
         private readonly ICRUD<Articelss> articelRepo;
         private readonly ApplicationDbContext db;
+        private readonly UserManager<ApplicationUser> userManger;
 
-        public DeleteController(ICRUD<Articelss> ArticelRepo, ApplicationDbContext Db)
+        public DeleteController(ICRUD<Articelss> ArticelRepo, ApplicationDbContext Db,UserManager<ApplicationUser> userManger)
         {
             articelRepo = ArticelRepo;
             db = Db;
+            this.userManger = userManger;
         }
 
         [HttpDelete]
         public IActionResult DeleteArticle(int id)
         {
-
-            articelRepo.Delete(id);
-            return Ok();
+            var aritcl = articelRepo.find(id);
+            var user = userManger.GetUserAsync(User).Result;
+            if (aritcl.ApplicationUser == user)
+            {
+                articelRepo.Delete(id);
+                return Ok();
+            }
+            else
+                return NotFound();
         }
     }
 }
